requirements.txt
scikit-learn
pandas
openai==1.1.2
python-dotenv==1.0.0